functions-python/run_post_match_tasks/requirements.txt
firebase-admin==5.2.0
python-dateutil
google-cloud-tasks==2.7.2